Add _compute_bounds() function to InfiniteLineVisual. Moved infinite_line.py to visuals folder.
