+++ vispy/visuals/__init__.py
@@ -19,12 +19,13 @@ from .gridlines import GridLinesVisual  # noqa
 from .image import ImageVisual  # noqa
 from .gridmesh import GridMeshVisual  # noqa
 from .histogram import HistogramVisual  # noqa
+from .infinite_line import InfiniteLineVisual  # noqa
 from .isocurve import IsocurveVisual  # noqa
 from .isoline import IsolineVisual  # noqa
 from .isosurface import IsosurfaceVisual  # noqa
-from .line import LineVisual, ArrowVisual, InfiniteLineVisual  # noqa
-from .line_plot import LinePlotVisual  # noqa
+from .line import LineVisual, ArrowVisual  # noqa
 from .linear_region import LinearRegionVisual  # noqa
+from .line_plot import LinePlotVisual  # noqa
 from .markers import MarkersVisual, marker_types  # noqa
 from .mesh import MeshVisual  # noqa
 from .plane import PlaneVisual  # noqa
