+++ vispy/visuals/infinite_line.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from ... import gloo
-from ..visual import Visual
+from .. import gloo
+from .visual import Visual
 
 
 VERT_SHADER = """
@@ -114,6 +114,19 @@ class InfiniteLineVisual(Visual):
             return self._pos[0, 0]
         else:
             return self._pos[0, 1]
+
+    def _compute_bounds(self, axis, view):
+        """Return the (min, max) bounding values of this visual along *axis*
+        in the local coordinate system.
+        """
+        is_vertical = self._is_vertical
+        pos = self._pos
+        if axis == 0 and is_vertical:
+            return (pos[0, 0], pos[0, 0])
+        elif axis == 1 and not is_vertical:
+            return (self._pos[0, 1], self._pos[0, 1])
+
+        return None
 
     @property
     def is_vertical(self):
